requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
selenium==4.37.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.2.0
